Initial commit
First draft of the LZSS decompression. So far tested only on the IMAGES.WDT, only on the header.

Don't pay too much attention to the current implementation/cleanliness - there many unknowns right now and so I expect code to change a lot. Method decomposition is on the TODO list but right now monolythic methods are easier to digest and debug.